Move old SLM modules into a separate package
