+++ src/mano-framework/plugins/service-lifecycle-manager/slm/old/helpers.py
